pusheos hasta ahora de posts

+++ backend/src/main/java/com/virtualwardrobe/backend/models/clothe/ClotheRepositorie.java
@@ -8,8 +8,10 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ClotheRepositorie extends JpaRepository<Clothe, Integer> {
+public interface ClotheRepositorie extends JpaRepository<Clothe, Long> {
+
     Optional<Clothe> findById(int id);
+
     @Query("SELECT DISTINCT c FROM Clothe c LEFT JOIN FETCH c.colorIds WHERE c.user.id = :userId")
     List<Clothe> findByUserId(int userId);
     // devuelve todas las prendas que tengan el id en la lista
